Simplify main file names within organized folders
- Rename request-schema.ts to schema.ts in request-schema folder
- Rename response-schema.ts to schema.ts in response-schema folder
- Rename request-validator.ts to validator.ts in request-validator folder
- Update all import paths for renamed files (12 files)
- Maintain clear structure: folder indicates purpose, file name is simple
- All changes compile successfully
