+++ packages/kori/src/request-validator/validate-params.ts
@@ -3,7 +3,7 @@ import { type KoriRequestSchemaDefault } from '../request-schema/index.js';
 import { ok, err, type KoriResult } from '../util/index.js';
 
 import { type KoriFieldValidationError } from './error.js';
-import { type KoriRequestValidatorDefault } from './request-validator.js';
+import { type KoriRequestValidatorDefault } from './validator.js';
 
 export async function validateRequestParams({
   validator,
