+++ packages/kori/src/request-validator/validated-request.ts
@@ -7,7 +7,7 @@ import {
   type KoriRequestSchemaDefault,
 } from '../request-schema/index.js';
 
-import { type KoriRequestValidatorDefault } from './request-validator.js';
+import { type KoriRequestValidatorDefault } from './validator.js';
 
 export type InferValidationOutput<S extends KoriRequestSchemaDefault> = {
   validatedParams(): InferRequestSchemaParamsOutput<S>;
